% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{create_lav_growth_model}
\alias{create_lav_growth_model}
\title{create_lav_growth_model}
\usage{
create_lav_growth_model(factor_name, manifest_vec, int.i.group_equal = F,
  var.i.group_equal = F, int.s.group_equal = F, var.s.group_equal = F,
  cov.group_equal = F, num_groups = 4, center_wave = 1)
}
\arguments{
\item{center_wave}{}
}
\description{
create_lav_growth_model
}
