% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{create_lav_id_strat_widaman}
\alias{create_lav_id_strat_widaman}
\title{create_lav_id_strat_widaman}
\usage{
create_lav_id_strat_widaman(factor_name)
}
\arguments{
\item{factor_name}{factor_name}
}
\description{
create_lav_id_strat_widaman
}
