% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{create_lav_ints}
\alias{create_lav_ints}
\title{create_lav_ints}
\usage{
create_lav_ints(manifest_vec, group_equal = F, long_free = F,
  num_groups = 4, fix_first_intercept = F)
}
\arguments{
\item{fix_first_intercept}{}
}
\description{
create_lav_ints
}
