% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{create_lav_invariance_model}
\alias{create_lav_invariance_model}
\title{create_lav_invariance_model}
\usage{
create_lav_invariance_model(factor_name, manifest_vec,
  type = c("unconstrained", "long_metric", "long_strong", "long_strict",
  "baseline", "metric", "strong", "strict"), num_groups = 4, no_widman = F,
  fix_first_intercept = F)
}
\arguments{
\item{factor_name}{factor_name}

\item{manifest_vec}{manifest_vec}

\item{type}{type}

\item{num_groups}{num_groups}

\item{no_widman}{no_widman}

\item{fix_first_intercept}{fix_first_intercept}
}
\description{
create_lav_invariance_model
}
