% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{create_lav_resid_covar}
\alias{create_lav_resid_covar}
\title{create_lav_resid_covar}
\usage{
create_lav_resid_covar(manifest_vec, group_equal = F, long_free = F,
  num_groups = 4)
}
\arguments{
\item{num_groups}{}
}
\description{
create_lav_resid_covar
}
