% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{create_lav_resid_var}
\alias{create_lav_resid_var}
\title{create_lav_resid_var}
\usage{
create_lav_resid_var(manifest_vec, group_equal = F, long_free = F,
  num_groups = 4)
}
\arguments{
\item{num_groups}{}
}
\description{
create_lav_resid_var
}
