% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{make_factor_string}
\alias{make_factor_string}
\title{make_factor_string}
\usage{
make_factor_string(items, factor_name)
}
\arguments{
\item{items}{items}

\item{factor_name}{factor_name}
}
\description{
make_factor_string
}
