% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mi_simulation_fit_data}
\alias{mi_simulation_fit_data}
\title{Measurement invariance simulation fit index data}
\format{A list of lists of matrices containing differences in fit statistics
  for constrained and unconstrained models.}
\usage{
mi_simulation_fit_data
}
\description{
1000 iterations per condition (loading strength by number of items per factor).
}
\keyword{datasets}
