% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{run_invariance_model}
\alias{run_invariance_model}
\title{run_invariance_model}
\usage{
run_invariance_model(factor_name, manifest_vec, item_data, type = "baseline",
  group = NULL, num_groups = 4, ...)
}
\arguments{
\item{...}{}
}
\description{
run_invariance_model
}
