% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{run_invariance_tests}
\alias{run_invariance_tests}
\title{run_invariance_tests}
\usage{
run_invariance_tests(factor_name, manifest_vec, item_data, group, ...)
}
\arguments{
\item{factor_name}{factor_name}

\item{manifest_vec}{manifest_vec}

\item{item_data}{item_data}

\item{group}{group}

\item{...}{...}
}
\description{
run_invariance_tests
}
