% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{test_factors_for_invariance}
\alias{test_factors_for_invariance}
\title{test_factors_for_invariance}
\usage{
test_factors_for_invariance(factor_names, indicators_df, item_data,
  fit_measures = c("mfi", "cfi", "rmsea", "rmsea.ci.lower", "rmsea.ci.upper"),
  group, save_fits = F, ...)
}
\arguments{
\item{factor_names}{factor_names}

\item{indicators_df}{indicators_df}

\item{item_data}{item_data}

\item{fit_measures}{fit_measures}

\item{group}{group}

\item{save_fits}{save_fits}

\item{...}{...}
}
\description{
test_factors_for_invariance
}
