% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{check_manifests_exist}
\alias{check_manifests_exist}
\title{check_manifests_exist}
\usage{
check_manifests_exist(manifest_vec, aDF)
}
\arguments{
\item{aDF}{}
}
\description{
check_manifests_exist
}
