% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{create_lav_factor_loadings}
\alias{create_lav_factor_loadings}
\title{create_lav_factor_loadings}
\usage{
create_lav_factor_loadings(factor_name, manifest_vec, group_equal = F,
  long_free = F, num_groups = 4)
}
\arguments{
\item{factor_name}{factor_name}

\item{manifest_vec}{manifest_vec}

\item{group_equal}{group_equal}

\item{long_free}{long_free}

\item{num_groups}{num_groups}
}
\description{
create_lav_factor_loadings
}
